% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prob_diff.R
\name{cal_prob_diff}
\alias{cal_prob_diff}
\title{The difference in probability}
\usage{
cal_prob_diff(df, center = FALSE)
}
\arguments{
\item{df}{A tibble contains 2 sets of probability and respective time grid}

\item{center}{Logical of whether to center the data or not}
}
\value{
A value of the test statistics and the associated p-value between two sets of real-time probabilistic forecasts
}
\description{
cal_prob_diff() returns the difference in probabilitistic forecasts of each instance at each time points
}
