% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eig.R
\name{calc_eig}
\alias{calc_eig}
\title{Calculator for eigenvalues}
\usage{
calc_eig(df, n_eig = 10, ngame, nsamp, grid = "grid", cent = FALSE)
}
\arguments{
\item{df}{A data frame that contains at least two columns, difference and the grid.}

\item{n_eig}{Number of leading eigenvalues to use}

\item{ngame}{Number of different instance}

\item{nsamp}{Number of sample points in the time domain}

\item{grid}{Name of the grid in the data frame}

\item{cent}{Whether to center the difference in probabilistic forecasts or not}
}
\value{
The leading \code{n_eig} eigenvalues of the covariance matrix to be used in the Delta test.
}
\description{
calc_eig() returns TBA
}
