% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_L_s2.R
\name{calc_L_s2}
\alias{calc_L_s2}
\title{Calculator for the Loss function and the corresponding variance}
\usage{
calc_L_s2(
  df,
  pA = "phat_1",
  pB = "phat_2",
  Y = "Y",
  grid = "grid",
  L = function(x, y) (x - y)^2
)
}
\arguments{
\item{df}{TBA}

\item{pA}{TBA}

\item{pB}{TBA}

\item{Y}{TBA}

\item{grid}{TBA}

\item{L}{TBA}
}
\value{
A value of the test statistics and the associated p-value between two sets of real-time probabilistic forecasts
}
\description{
calc_L_s2() returns TBA
}
