% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pval.R
\name{calc_pval}
\alias{calc_pval}
\title{Calculator of the p-values of the delta test}
\usage{
calc_pval(Z, eig, quan, n_MC = 5000)
}
\arguments{
\item{Z}{TBA}

\item{eig}{TBA}

\item{quan}{TBA}

\item{n_MC}{TBA}
}
\value{
A value of the test statistics and the associated p-value between two sets of real-time probabilistic forecasts
}
\description{
calc_pval() returns TBA
}
