% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Z.R
\name{calc_Z}
\alias{calc_Z}
\title{Calculator of the test statistics of delta test}
\usage{
calc_Z(
  df,
  pA,
  pB,
  Y,
  grid = "grid",
  nsamp,
  ngame,
  L = function(x, y) (x - y)^2
)
}
\arguments{
\item{df}{A data frame that contains the following columns: Y, pA, pB, and grid}

\item{pA}{Name of the probabilistic forecasts from the first method}

\item{pB}{Name of the probabilistic forecasts from the second method}

\item{Y}{Name of the outcome variable}

\item{grid}{Name of the grid}

\item{nsamp}{Number of sample points in the time domain}

\item{ngame}{Number of different instance}

\item{L}{Loss function to evaluate the performance between two methods}
}
\value{
A value of the test statistics between two sets of real-time probabilistic forecasts
}
\description{
calc_Z() returns a value of the test statistics of the delta test.
}
