% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_gen.R
\name{df_gen}
\alias{df_gen}
\title{TBA}
\usage{
df_gen(N, Ngame, type = "OU", a = 1, b = 0.27)
}
\arguments{
\item{N}{TBA}

\item{Ngame}{TBA}

\item{type}{TBA}

\item{a}{TBA}

\item{b}{TBA}
}
\value{
TBA
}
\description{
df_gen() TBA
}
