% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalProbForecast-package.R
\docType{package}
\name{evalProbForecast-package}
\alias{evalProbForecast}
\alias{evalProbForecast-package}
\title{evalProbForecast: Evaluate the performance between two probabilistic forecasts}
\description{
This package provides the tools to evaluate the performance between two sets of probabilistic forecasts, under the square norm, along with the graphical representation.
}
\author{
\strong{Maintainer}: Chi-Kuang Yeh \email{chi-kuang.yeh@uwaterloo.ca}

Other contributors:
\itemize{
  \item Gregory Rice [contributor, thesis advisor]
  \item Joel A. Dubin [contributor, thesis advisor]
}

}
\keyword{internal}
