% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalRTPF-package.R
\docType{package}
\name{evalRTPF-package}
\alias{evalRTPF}
\alias{evalRTPF-package}
\title{evalRTPF: Evaluate the discrepancy between two Real-Time updated Probabilistic Forecasts}
\description{
This package provides the evaluating tools to on the performance between two or more competing continuously-updated probabilistic forecasts, under the square norm, along with an easy-to-interpret graphical representation.
}
\author{
\strong{Maintainer}: Chi-Kuang Yeh \email{chi-kuang.yeh@uwaterloo.ca}

Other contributors:
\itemize{
  \item Gregory Rice [contributor, thesis advisor]
  \item Joel A. Dubin [contributor, thesis advisor]
}

}
\keyword{internal}
