% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_interp.R
\name{lin_interp}
\alias{lin_interp}
\title{Linear Interpolation}
\usage{
lin_interp(prob, grid, outcome)
}
\arguments{
\item{prob}{TBA}

\item{grid}{TBA}

\item{outcome}{TBA}
}
\value{
A tibble contains a set of interpolated probabilistic forecasts, outcome variable and grid.
}
\description{
lin_interp() returns the set of input probabilistic forecasts evaluated on the given grid
}
