% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pcb.R
\name{plot_pcb}
\alias{plot_pcb}
\title{Plotter for point-wise confidence band}
\usage{
plot_pcb(df, grid = "grid", L = "L", var = "sigma2")
}
\arguments{
\item{df}{TBA}

\item{grid}{TBA}

\item{L}{TBA}

\item{var}{TBA}
}
\value{
A value of the test statistics and the associated p-value between two sets of real-time probabilistic forecasts
}
\description{
plot_pcb() returns TBA
}
