% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eig.R
\name{calc_eig}
\alias{calc_eig}
\title{Calculator for eigenvalues}
\usage{
calc_eig(df, n_eig = 10, ngame, nsamp, grid = "grid", cent = FALSE)
}
\arguments{
\item{df}{TBA}

\item{n_eig}{TBA}

\item{ngame}{TBA}

\item{nsamp}{TBA}

\item{grid}{TBA}

\item{cent}{A tibble contains 2 sets of probability and respective time grid}
}
\value{
A value of the test statistics and the associated p-value between two sets of real-time probabilistic forecasts
}
\description{
calc_eig() returns TBA
}
