% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eig.R
\name{calc_eig}
\alias{calc_eig}
\title{TBA}
\usage{
calc_eig(cent = FALSE, df, n_eig = 10, ngame, nsamp)
}
\arguments{
\item{cent}{A tibble contains 2 sets of probability and respective time grid}

\item{df}{TBA}

\item{n_eig}{TBA}

\item{ngame}{TBA}

\item{nsamp}{TBA}
}
\value{
A value of the test statistics and the associated p-value between two sets of real-time probabilistic forecasts
}
\description{
calc_eig() returns TBA
}
