% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Z.R
\name{calc_Z}
\alias{calc_Z}
\title{Calculator of the test statistics of delta test}
\usage{
calc_Z(df, pA, pB, Y, grid, nsamp, ngame, L = function(x, y) (x - y)^2)
}
\arguments{
\item{df}{A tibble contains 2 sets of probability and respective time grid}

\item{pA}{TBA}

\item{pB}{TBA}

\item{Y}{TBA}

\item{grid}{TBA}

\item{nsamp}{TBA}

\item{ngame}{TBA}

\item{L}{Logical of whether to centre the data or not}
}
\value{
A value of the test statistics and the associated p-value between two sets of real-time probabilistic forecasts
}
\description{
calc_Z() returns TBA
}
