% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_test.R
\name{delta_test}
\alias{delta_test}
\title{A statistical test based on the square norm}
\usage{
delta_test(
  df,
  D = 10,
  N_MC = 5000,
  center = TRUE,
  grid = "grid",
  pA = "phat_A",
  pB = "phat_B",
  L = function(x, y) {     (x - y)^2 }
)
}
\arguments{
\item{df}{First set of probabilistic forecasts}

\item{D}{Scalar of how many eigenvalues to use}

\item{N_MC}{Scalar of how many Monte Carlo simulation to do}

\item{center}{Logical of whether to center the data or not}

\item{grid}{Logical of whether to center the data or not}

\item{pA}{Logical of whether to center the data or not}

\item{pB}{Logical of whether to center the data or not}

\item{L}{Logical of whether to center the data or not}
}
\value{
A value of the test statistics and the associated p-value between two sets of real-time probabilistic forecasts
}
\description{
delta_test() returns the test statistics between two real-time probabilistic forecasts under square norm
}
